Correct SIMM sample plugin
* Split SIMM plugin into separate plugin registry extension and web server plugin to resolve class loading error.
* Rationalise differences between Bank of Corda and SIMM demo build scripts to make side by side comparison easier.
