+++ samples/bank-of-corda-demo/build.gradle
@@ -23,7 +23,6 @@ configurations {
 
 dependencies {
     compile "org.jetbrains.kotlin:kotlin-stdlib-jre8:$kotlin_version"
-    testCompile "junit:junit:$junit_version"
 
     // Corda integration dependencies
     cordaCompile project(path: ":node:capsule", configuration: 'runtimeArtifacts')
@@ -37,6 +36,12 @@ dependencies {
 
     // Javax is required for webapis
     compile "org.glassfish.jersey.core:jersey-server:${jersey_version}"
+
+    // Cordapp dependencies
+    // Specify your cordapp's dependencies below, including dependent cordapps
+
+    // Test dependencies
+    testCompile "junit:junit:$junit_version"
 }
 
 task deployNodes(type: net.corda.plugins.Cordform, dependsOn: ['jar']) {
