+++ samples/simm-valuation-demo/src/main/resources/META-INF/services/net.corda.core.node.CordaPluginRegistry
@@ -1,2 +1,2 @@
 # Register a ServiceLoader service extending from net.corda.core.node.CordaPluginRegistry
-net.corda.vega.services.SimmService$Plugin
+net.corda.vega.plugin.SimmPluginRegistry
