+++ samples/simm-valuation-demo/src/main/resources/META-INF/services/net.corda.webserver.services.WebServerPluginRegistry
@@ -1,2 +1,2 @@
 # Register a ServiceLoader service extending from net.corda.webserver.services.WebServerPluginRegistry
-net.corda.vega.services.SimmService$Plugin
+net.corda.vega.plugin.SimmPlugin
